chore: use CurrentUserDetails in hibernate store class
