+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/user/CurrentUserDetails.java
@@ -30,6 +30,7 @@ package org.hisp.dhis.user;
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.Map;
+import java.util.Set;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -58,7 +59,16 @@ public interface CurrentUserDetails extends UserDetails
     @Override
     boolean isEnabled();
 
+    boolean isSuper();
+
     String getUid();
+
+    /**
+     * Set of UserGroup UID which current User belongs to.
+     *
+     * @return
+     */
+    Set<String> getUserGroupIds();
 
     Map<String, Serializable> getUserSettings();
 }
