+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/user/DefaultUserService.java
@@ -839,6 +839,8 @@ public class DefaultUserService
             .credentialsNonExpired( credentialsNonExpired )
             .authorities( user.getAuthorities() )
             .userSettings( new HashMap<>() )
+            .userGroupIds( currentUserService.getCurrentUserGroupsInfo( user ).getUserGroupUIDs() )
+            .isSuper( user.isSuper() )
             .build();
     }
 }
